requirements.txt
PySide6
Pillow
nuitka
zstandard
